Change Shares Archive Key to bigint

+++ sql/000_base_structure.sql
@@ -155,7 +155,7 @@ CREATE TABLE IF NOT EXISTS `shares` (
 ) ENGINE=InnoDB  DEFAULT CHARSET=utf8;
 
 CREATE TABLE IF NOT EXISTS `shares_archive` (
-  `id` int(255) unsigned NOT NULL AUTO_INCREMENT,
+  `id` bigint(30) unsigned NOT NULL AUTO_INCREMENT,
   `share_id` int(255) unsigned NOT NULL,
   `username` varchar(120) NOT NULL,
   `our_result` enum('Y','N') DEFAULT NULL,
